fix: consolidate local-midnight DST math onto lib/timezone.js
server/lib/timezone.js and server/services/humanActivity.js each had
their own UTC-instant-of-local-midnight implementation, with different
DST-correction strategies — one refines the sampled UTC offset at the
candidate instant, the other checks the landed-vs-desired delta via
getLocalParts. Both were correct on their current inputs, but forked
DST math invites drift.

lib/timezone.js is now the single owner:
- localDayRangeUtc(dateStr, timezone) is ported into the lib, built on
  the existing anchorLocalMidnightUtc helper, as a new export
  alongside (not merged into) localDayWindowUtc — the two have a real
  semantic difference (exclusive next-midnight end vs. inclusive
  +24h-1ms) that serves different callers.
- humanActivity.js's private localMidnightUtc is deleted; its own
  resolveEventInstant() now calls anchorLocalMidnightUtc directly.
- Its documented caveat about getUtcOffsetMs mis-parsing ambiguous
  wall-clock times at a DST transition moves into the lib's JSDoc,
  where the two-pass refinement that caveat is about actually lives.
- youtubeSync.js and goalScorecard.js now import localDayRangeUtc from
  the lib directly instead of via humanActivity.js.

The DST-transition parity tests (23h spring-forward / 25h fall-back
days, month/year rollover) move from humanActivity.test.js into
timezone.test.js, verifying the ported implementation reproduces the
exact same instants the deleted landed-vs-desired strategy did —
confirmed by running both the moved suite and every consumer's own
test file (129 tests, all green) against the real change, not just
the diff.

Also expanded lib/README.md's timezone.js row to actually document
its exports (was a one-line stub) per the acceptance criteria.

Closes #5572
