+++ server/lib/timezone.test.js
@@ -241,6 +241,12 @@ describe('timezone', () => {
   })
 
   describe('localDayRangeUtc', () => {
+    it('returns a 24h UTC window for a constant-offset local day', () => {
+      const range = localDayRangeUtc('2026-07-04', 'America/Los_Angeles')
+      expect(range.start.toISOString()).toBe('2026-07-04T07:00:00.000Z')
+      expect(range.end.getTime() - range.start.getTime()).toBe(24 * 60 * 60 * 1000)
+    })
+
     it.each([
       ['2026-03-08', '2026-03-08T08:00:00.000Z', '2026-03-09T07:00:00.000Z', 23],
       ['2026-11-01', '2026-11-01T07:00:00.000Z', '2026-11-02T08:00:00.000Z', 25],
@@ -259,6 +265,13 @@ describe('timezone', () => {
       const range = localDayRangeUtc(' 2026-03-08 ', 'America/Los_Angeles')
       expect(range.start.toISOString()).toBe('2026-03-08T08:00:00.000Z')
       expect(range.end.toISOString()).toBe('2026-03-09T07:00:00.000Z')
+    })
+
+    it('rolls the end boundary over month and year edges', () => {
+      const monthEdge = localDayRangeUtc('2026-01-31', 'UTC')
+      expect(monthEdge.end.toISOString()).toBe('2026-02-01T00:00:00.000Z')
+      const yearEdge = localDayRangeUtc('2026-12-31', 'UTC')
+      expect(yearEdge.end.toISOString()).toBe('2027-01-01T00:00:00.000Z')
     })
   })
